Use double quotes for non-system libraries.
The idea here is that any header file that is part of this project,
including any submodule, is not a system library - and therefore
should be included using double quotes.

Rationale: although the C standard doesn't specify where these header
files are searched for (it says it is implementation-defined), it
does say that files included using double quotes are searched for one
way, and when that fails, then they are search for as if included in
angle brackets.

This allows one to avoid finding including headers that might not
be intended because they aren't on the developers machine, for example.
Most notably (in any sane compiler implementation): if a file
foo.c contains an #include "foo.h" then the compiler will FIRST
look for foo.h in the same directory that foo.c resides in, and
only then elsewhere. But provisions exist (at least for the GNU
compiler) to also make sure that - if this fails - to first look in
places that are part of the project BEFORE looking elsewhere on
the system (aka /usr/include et al). Currently px4 doesn't use these
provisions, so that there is no difference between "" and <>
unless the header file resides in the same directory as the file
including it. But, it seems to be a good programming style to use
double quotes here: the only reason that this commit (and future
ones that I planned) would be wrong is when it is the explicit desire
that libraries installed on the system should have PRECEDENCE over
(say) submodules libraries; but considering the SHA-1 hash that is
used to link precise versions of the submodules to any given commit
of master I doubt that this is the case.

This is the first of several commits that prepare for a project-wide
fix of this issue by means of a new script 'Tools/fix_headers.sh' (to
be committed) which will not touch the include in THIS commit because
it is too deep (too far away from the top of the file). Hence the need
for this manual commit for this case.